4.3.0-0.1.0: [feature] - Multichain SDK (#192)
* WIP

* Refactor

* Fix missing event error

* Expose observable rather than subject

* Fix circular dependency

* Add docs

* Add timeout for unsub

* throw error if unsupported network

* Fix network check

* Increment version

* Remove unused import
