+++ src/event.ts
@@ -1,6 +1,7 @@
-import { EventObject } from './interfaces'
+import SDK from '.'
+import { EventObject } from './types'
 
-function event(this: any, eventObj: EventObject): void {
+function event(this: SDK, eventObj: EventObject): void {
   if (this._destroyed)
     throw new Error(
       'The WebSocket instance has been destroyed, re-initialize to continue making requests.'
