+++ src/streams.ts
@@ -1,4 +1,4 @@
 import { Subject } from 'rxjs'
-import { EthereumTransactionData, SimulationTransactionOutput, SimulationTransaction } from './interfaces'
+import { SimulationTransactionOutput } from './types'
 
 export const simulations$ = new Subject<SimulationTransactionOutput>()
